Implemented most of the app
- Lifted most of the CSS from davidkpiano's example app
- Added a few react components, it felt right

The app consists of the search form and the results panel and a zoomed
in photo.

The events generally bubble up to the top which holds the most of the
state.

The search action has a long delay to fake cancelability, since the
fetch APIs don't generally support cancels.

The code _works_ but the behaviour is quite buggy, partly because of
the bad design, the dependency on booleans and "forgetting" to reset
them when they should be reset.